requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
langchain==0.0.340
langchain-google-genai==1.0.1
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
scikit-learn==1.3.2
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0